requirements.txt
adatest==0.3.5
aiohttp==3.9.3
aiohttp-security==0.5.0
aiohttp-session==2.12.0
aiosignal==1.3.1
altair==5.1.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cryptography==42.0.5
cycler==0.11.0
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nba-api==1.2.1
nest-asyncio==1.5.8
networkx==3.1
numba==0.59.0
numpy==1.26.0
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.9.1
platformdirs==3.10.0
profanity==1.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.4.1.post1
scipy==1.12.0
shap==0.44.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
utils==1.0.1
validators==0.22.0
virtualenv==20.24.4
watchdog==3.0.0
wcwidth==0.2.6
whitenoise==6.5.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0